requirements.txt
fastapi
uvicorn[standard]
SQLAlchemy==2.0.19
pydantic
requests==2.31.0
python-dotenv==1.0.0
aiofiles==23.1.0
jinja2==3.1.2
pytest==7.4.2
dicttoxml==1.7.4
reportlab==4.1.0
python-dateutil==2.8.2

